pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vts-gui"
version = "1.2.0"
description = "VTS GUI: A Whisper-based transcription and translation tool for subtitles"
authors = [
  { name="Josiah Watkins", email="veilbreaker@voidmaster.xyz" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9.9"
dependencies = [
  "torch==2.5.1+cu121",
  "openai-whisper @ git+https://github.com/openai/whisper.git"
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]

[project.urls]
"Homepage" = "https://github.com/menevia16a/VTS-GUI"
"Source" = "https://github.com/menevia16a/VTS-GUI"
"Tracker" = "https://github.com/menevia16a/VTS-GUI/issues"

[project.scripts]
vts-gui = "vts_gui.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["vts_gui*"]
exclude = ["Media*", "Output*"]
